Update parser.js
The subsets in EDAM have changed since 1.25, changing from `http://purl.obolibrary.org/obo/edam#topics` to `http://edamontology.org/topics`. This causes them to be interpreted as top-level classes by the parser, and these subsets (only the ones for topics, operations, data and formats) to be displayed in EDAM browser. This correction allows to avoid this problem by restricting the control on the top-level superclass IDs.